7.3.0

+++ RTIMULib/RunningAverage.cpp
@@ -18,7 +18,7 @@
 // 0.2.03 - 2013-11-31 getElement
 // 0.2.04 - 2014-07-03 added memory protection
 // 0.2.05 - 2014-12-16 changed float -> float
-// 0.2.06 - 2015-03-07 all size uint8_t
+// 0.2.06 - 2015-03-07 all size uint16_t
 // 0.2.07 - 2015-03-16 added getMin() and getMax() functions (Eric Mulder)
 // 0.2.08 - 2015-04-10 refactored getMin() and getMax() implementation
 //
@@ -28,7 +28,7 @@
 #include "RunningAverage.h"
 #include <stdlib.h>
 
-RunningAverage::RunningAverage(uint8_t size)
+RunningAverage::RunningAverage(uint16_t size)
 {
     _size = size;
     _ar = (float*) malloc(_size * sizeof(float));
@@ -49,7 +49,7 @@ void RunningAverage::clear()
     _sum = 0.0;
     _min = NAN;
     _max = NAN;
-    for (uint8_t i = 0; i < _size; i++)
+    for (uint16_t i = 0; i < _size; i++)
     {
         _ar[i] = 0.0f; // keeps addValue simple
     }
@@ -80,7 +80,7 @@ float RunningAverage::getAverage()
 }
 
 // returns the value of an element if exist, NAN otherwise
-float RunningAverage::getElement(uint8_t idx)
+float RunningAverage::getElement(uint16_t idx)
 {
     if (idx >=_cnt ) return NAN;
     return _ar[idx];
@@ -89,11 +89,11 @@ float RunningAverage::getElement(uint8_t idx)
 // fill the average with a value
 // the param number determines how often value is added (weight)
 // number should preferably be between 1 and size
-void RunningAverage::fillValue(float value, uint8_t number)
+void RunningAverage::fillValue(float value, uint16_t number)
 {
     clear(); // TODO conditional?  if (clr) clear();
 
-    for (uint8_t i = 0; i < number; i++)
+    for (uint16_t i = 0; i < number; i++)
     {
         addValue(value);
     }
